requirements.txt
# Core dependencies for file processing and data matching
pandas>=2.0.0
numpy>=1.24.0
rapidfuzz>=3.0.0
tqdm>=4.65.0

# Configuration and validation
pydantic>=2.0.0
jsonschema>=4.17.0
pyyaml>=6.0
watchdog>=3.0.0

# Web framework
flask>=2.3.0
werkzeug>=2.3.0
flask-socketio>=5.3.0
websockets>=11.0.0
flask-jwt-extended>=4.5.0
flask-restx>=1.1.0
flask-cors>=4.0.0
flask-limiter>=3.3.0

# Async and concurrency
asyncio-throttle>=1.0.2

# File format support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Text processing and normalization
unidecode>=1.3.6
regex>=2023.6.3

# Caching and performance
redis>=4.5.0
lru-dict>=1.2.0

# Logging and monitoring
structlog>=23.1.0
prometheus-client>=0.16.0
psutil>=5.9.0

# Development and testing (optional)
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Type hints
types-redis>=4.5.0
types-pyyaml>=6.0.0